refactor: organize enigma helper methods into appropriately grouped modules.
